Рефакторит утилитарные функции
